Add files via upload
The T-12-images are reflected light images of the interior of a Mn-nodule (from the Pacific Ocean). The nodule itself consists of several smaller nodules, grown together.  Clearly identifiable crystals of pyrolusite (pleochroitic, with cleavage) are formed between interior nodules consisting of aggregates of cryptomelane/romanechite. These aggregates are cryptocrystalline and the randomly oriented crystals of cryptomelane/romanechite may lead to the aggregates appearing isotropic (which they are not). The cryptocrystalline aggregates may be mistaken for tennantite (Cu6(Cu4C2+2)As4S12S). However, tennantite does not occur in this chemical environment. Cryptomelane/romanechite cannot be optically identified in such aggregates. This should be dome with e.g. X-Ray Diffraction.