Add files via upload
Cryptocrystalline aggregates of cryptomelane/romanechite with pyrolusite crystals. The cryptomelane/romanechite aggregates may appear isotropic (but they are NOT), due to the extremely small crystal sizes. Optical identification of such an aggregates can easily yield tennantite (Cu6(Cu4C2+2)As4S12S) as a solution but this is a mineral that does not occur in this chemical system. Texture and chemical system are also important in identification of minerals!!